Fixed a major problem with edit function
If there were same names in 2 or more contacts then the edit function would get confused in which of the contacts would it edit and by default it would edit the first one. This was fixed by using the number if two or more contacts were found. The add function was also modified in order to accept a unique number for every contact to avoid any further problems.